requirements.txt
Python 3.9.6
Flask==2.2.3
requests==2.28.2
pyspellchecker==0.7.2
nltk==3.8.1
numpy==1.23.5
torch==2.0.0
transformers==4.25.1
scikit-learn==1.2.2
beautifulsoup4==4.12.0
docker==6.0.1
farm-haystack==1.15.1
